update tests, split tests into executor_test.go
