+++ pkg/compactor/executor.go
@@ -1,7 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-only
-// Provenance-includes-location: https://github.com/cortexproject/cortex/blob/master/pkg/compactor/compactor.go
-// Provenance-includes-license: Apache-2.0
-// Provenance-includes-copyright: The Cortex Authors.
 
 package compactor
 
@@ -22,12 +19,9 @@ import (
 	"github.com/grafana/mimir/pkg/util"
 )
 
-// CompactionExecutor defines how compaction work gets executed.
+// CompactionExecutor defines how compaction work is executed.
 type CompactionExecutor interface {
-	// Run executes the compaction strategy. This method should block until
-	// the context is cancelled or an error occurs.
 	Run(ctx context.Context, compactor *MultitenantCompactor) error
-	// Stop performs cleanup when the executor is being shut down.
 	Stop() error
 }
 
@@ -54,13 +48,12 @@ func (e *StandaloneExecutor) Run(ctx context.Context, c *MultitenantCompactor) e
 	}
 }
 
-// Stop implements CompactionExecutor for standalone mode.
+// Stop implements CompactionExecutor
 func (e *StandaloneExecutor) Stop() error {
-	// StandaloneExecutor has no resources to clean up
 	return nil
 }
 
-// SchedulerExecutor requests compaction jobs from a scheduler service.
+// SchedulerExecutor requests compaction jobs from an external scheduler.
 type SchedulerExecutor struct {
 	cfg                      Config
 	logger                   log.Logger
@@ -77,22 +70,20 @@ func NewSchedulerExecutor(cfg Config, logger log.Logger, invalidClusterValidatio
 		invalidClusterValidation: invalidClusterValidation,
 	}
 
-	// Initialize scheduler client - this will succeed even if scheduler is unreachable
-	// since grpc.Dial() creates the connection object immediately and connects lazily
+	// Initialize scheduler client. This call will succeed even if scheduler is unreachable
+	// since grpc.Dial() creates the connection immediately and connects lazily.
 	var err error
 	executor.schedulerClient, executor.schedulerConn, err = executor.makeSchedulerClient()
 	if err != nil {
-		return nil, errors.Wrap(err, "failed to create scheduler client")
+		return nil, err
 	}
 
 	return executor, nil
 }
 
-// Run implements CompactionExecutor for scheduler mode.
+// Run implements CompactionExecutor
 func (e *SchedulerExecutor) Run(ctx context.Context, c *MultitenantCompactor) error {
-	// Scheduler client is already initialized in NewSchedulerExecutor
 
-	// Generate a unique worker ID for this compactor instance
 	workerID := fmt.Sprintf("compactor-%s", c.ringLifecycler.GetInstanceID())
 	level.Info(e.logger).Log("msg", "compactor running in worker mode", "scheduler_endpoint", e.cfg.SchedulerAddress, "worker_id", workerID)
 
@@ -120,6 +111,14 @@ func (e *SchedulerExecutor) Run(ctx context.Context, c *MultitenantCompactor) er
 			return errors.Wrap(err, "compactor subservice failed")
 		}
 	}
+}
+
+// Stop implements CompactionExecutor
+func (e *SchedulerExecutor) Stop() error {
+	if e.schedulerConn != nil {
+		return e.schedulerConn.Close()
+	}
+	return nil
 }
 
 func (e *SchedulerExecutor) makeSchedulerClient() (schedulerpb.CompactorSchedulerClient, *grpc.ClientConn, error) {
@@ -157,51 +156,43 @@ func (e *SchedulerExecutor) leaseAndExecuteJob(ctx context.Context, c *Multitena
 		level.Warn(e.logger).Log("msg", "failed to mark job as in progress", "job_id", resp.Key.Id, "err", err)
 	}
 
-	// Create context with timeout for job execution based on max job duration
 	jobCtx, jobCancel := context.WithTimeout(ctx, e.cfg.SchedulerMaxJobDuration)
 	defer jobCancel()
+	statusChan := make(chan schedulerpb.UpdateType, 1)
 
-	// Create done channel to signal job completion to status update goroutine
-	done := make(chan struct{})
-	defer close(done)
-
-	// Start per-job status update goroutine
+	// Status updater goroutine handles both keep-alive status updates and reports final job status
 	go func() {
 		ticker := time.NewTicker(e.cfg.SchedulerUpdateInterval)
 		defer ticker.Stop()
 
 		for {
 			select {
-			case <-done:
-				return // Job completed, stop sending updates
 			case <-ticker.C:
 				if err := e.updateJobStatus(ctx, resp.Key, resp.Spec.Tenant, schedulerpb.IN_PROGRESS); err != nil {
-					level.Warn(e.logger).Log("msg", "failed to send keep-alive status update", "job_id", resp.Key.Id, "err", err)
+					level.Warn(e.logger).Log("msg", "failed to send status update", "job_id", resp.Key.Id, "tenant", resp.Spec.Tenant, "status", schedulerpb.IN_PROGRESS, "err", err)
 				}
+			case status := <-statusChan:
+				// TODO: Consider retry logic for status update failures. Currently we only log a warning
+				// if the scheduler is unreachable during the final status udpate.
+				if err := e.updateJobStatus(ctx, resp.Key, resp.Spec.Tenant, status); err != nil {
+					level.Warn(e.logger).Log("msg", "failed to send final status update", "job_id", resp.Key.Id, "tenant", resp.Spec.Tenant, "status", status, "err", err)
+				}
+				return
 			case <-ctx.Done():
 				return
 			}
 		}
 	}()
 
-	// Execute the actual compaction job with timeout context
-	updType, err := e.executeCompactionJob(jobCtx, c, resp.Spec)
+	status, err := e.executeCompactionJob(jobCtx, c, resp.Spec)
 	if err != nil {
-		// Check if the error is due to job timeout
 		if errors.Is(err, context.DeadlineExceeded) {
 			level.Warn(e.logger).Log("msg", "compaction job exceeded max duration", "job_id", resp.Key.Id, "tenant", resp.Spec.Tenant, "max_duration", e.cfg.SchedulerMaxJobDuration)
-			// Send failure status for timeout
-			if statusErr := e.updateJobStatus(ctx, resp.Key, resp.Spec.Tenant, schedulerpb.ABANDON); statusErr != nil {
-				level.Warn(e.logger).Log("msg", "failed to update job status after timeout", "job_id", resp.Key.Id, "err", statusErr)
-			}
 		}
+		statusChan <- schedulerpb.ABANDON
 		return true, err
 	}
-
-	// Send final status update
-	if err := e.updateJobStatus(ctx, resp.Key, resp.Spec.Tenant, updType); err != nil {
-		return true, err
-	}
+	statusChan <- status
 	return true, nil
 }
 
@@ -211,15 +202,7 @@ func (e *SchedulerExecutor) updateJobStatus(ctx context.Context, key *schedulerp
 	return err
 }
 
+// executeCompactionJob executes the compaction work for a scheduled job.
 func (e *SchedulerExecutor) executeCompactionJob(ctx context.Context, c *MultitenantCompactor, spec *schedulerpb.JobSpec) (schedulerpb.UpdateType, error) {
-	level.Info(e.logger).Log("msg", "executing compaction job", "tenant", spec.Tenant)
 	return schedulerpb.COMPLETE, nil
-}
-
-// Stop implements CompactionExecutor for scheduler mode.
-func (e *SchedulerExecutor) Stop() error {
-	if e.schedulerConn != nil {
-		return e.schedulerConn.Close()
-	}
-	return nil
 }
